Move test event handler to test package
